App/requirements.txt
streamlit
pandas
python-nmap
shodan
matplotlib
sqlalchemy